Add files via upload
"FRR" or Flight Readiness Review is the final document sent to NASA in the USLI competition.  This is the culmination of near ten thousand hours of work between myself and many other individuals who worked on this project. I saw this project begin and end, and had a hand in nearly every page on this document.